pyproject.toml
[project]
name = "noai"
version = "0.1.0"
description = "My AI coding agent"
authors = [{name = "Noah Toral", email = "arknt8099@gmail.com"}]
requires-python = ">=3.8"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"